Esercizio completo con RepoMock

+++ Week7.Master.Core/BusinessLayer/MainBusinessLayer.cs
@@ -67,15 +67,25 @@ namespace Week7.Master.Core.BusinessLayer
                 return "Errore: Codice errato.";
             }
 
-            //TODO:non deve essere possibile cancellare un corso che ha almeno una lezione associata
+            //non deve essere possibile cancellare un corso che ha almeno una lezione associata
+            var lezioneAssociataAlCorso = GetAllLezioni().FirstOrDefault(l => l.CorsoCodice == corsoEsistente.CorsoCodice);
+            if (lezioneAssociataAlCorso != null)
+            {
+                return "Impossibile cancellare il corso perchè risulta associato ad almeno una lezione";
+            }
             //nè un corso che ha almeno uno studente iscritto.
-
+            var studenteAssociataAlCorso = GetAllStudenti().FirstOrDefault(s => s.CorsoCodice == corsoEsistente.CorsoCodice);
+            if (studenteAssociataAlCorso != null)
+            {
+                return "Impossibile cancellare il corso perchè risulta associato ad almeno uno studente";
+            }
             corsiRepo.Delete(corsoEsistente);
             return "Corso eliminato correttamente";
 
         }
 
         #endregion
+
 
         #region funzionalità Studenti
         public List<Studente> GetAllStudenti()
@@ -131,6 +141,141 @@ namespace Week7.Master.Core.BusinessLayer
                 return null;
             }            
             return studentiRepo.GetAll().Where(s => s.CorsoCodice == corso.CorsoCodice).ToList();
+        }
+        #endregion
+
+
+        #region funzionalità Docenti
+        public IList<Docente> GetAllDocenti()
+        {
+            return docentiRepo.GetAll().ToList();
+        }
+        public string InserisciNuovoDocente(Docente nuovoDocente)
+        {
+            //controllo input
+            //non deve esistere un altro docente con lo stesso nome, cognome e email
+            var docenteEsistente = GetAllDocenti().FirstOrDefault(d => d.Nome == nuovoDocente.Nome && d.Cognome == nuovoDocente.Cognome && d.Email == nuovoDocente.Email);
+
+            if (docenteEsistente != null)
+            {
+                return "Docente già esistente";
+            }
+            docentiRepo.Add(nuovoDocente);
+            return "Docente Aggiunto con successo";
+        }
+        public string ModificaDocente(int id, string nuovoTelefono, string nuovaMail)
+        {
+            //recupero il docente tramite l'id e modifico le 2 proprietà
+            var docenteDaModificare = docentiRepo.GetAll().FirstOrDefault(d => d.ID == id);
+            if (docenteDaModificare == null)
+            {
+                return "Id non valido/ non presente";
+            }
+            docenteDaModificare.Telefono = nuovoTelefono;
+            docenteDaModificare.Email = nuovaMail;
+            docentiRepo.Update(docenteDaModificare);
+            return "Modifica effettuata";
+        }
+
+        public string EliminaDocente(int idDocenteDaEliminare)
+        {
+            //controllo su input
+            var docenteEsistente = docentiRepo.GetById(idDocenteDaEliminare);
+            if (docenteEsistente == null)
+            {
+                return "Id non valido.Impossibile eliminare.";
+            }
+            var docenteAssociatoALezione = GetAllLezioni().FirstOrDefault(l => l.DocenteID == idDocenteDaEliminare);
+            if (docenteAssociatoALezione != null)
+            {
+                return "Impossibile cancellare il docente perchè risulta associato ad almeno una lezione";
+            }
+            docentiRepo.Delete(docenteEsistente);
+            return "Docente eliminato con successo";
+        }
+        #endregion
+
+
+        #region Funzionalità Lezioni
+        public IList<Lezione> GetAllLezioni()
+        {
+            return lezioniRepo.GetAll().ToList();
+        }
+
+        public string AggiungiLezione(Lezione lezione)
+        {
+            //controllo input
+            //controllo se codice corso esiste
+            var corso = corsiRepo.GetByCode(lezione.CorsoCodice);
+            if (corso == null)
+            {
+                return "Codice Corso errato o inesistente";
+            }
+
+            //e se esiste codice docente
+            var docente = docentiRepo.GetById(lezione.DocenteID);
+            if (docente == null)
+            {
+                return "Codice docente inesistente";
+            }
+            //Si possono eventualmente prevedere altri controlli ad esempio verifica che non esista già
+            //una lezione associata allo stesso docente lo stesso giorno..
+
+            lezioniRepo.Add(lezione);
+            return "Aggiunta correttamente";
+        }
+
+        public string ModificaLezione(int id, string nuovaAula)
+        {
+            //recupero la Lezione tramite l'id 
+            var lezioneDaModificare = lezioniRepo.GetAll().FirstOrDefault(l => l.LezioneID == id);
+            if (lezioneDaModificare == null)
+            {
+                return "Id non valido/ non presente";
+            }
+            //Modifico il campo Aula con il nuovo valore scritto dall'utente
+            lezioneDaModificare.Aula = nuovaAula;
+            //aggiorno la lezione
+            lezioniRepo.Update(lezioneDaModificare);
+            return "Modifica effettuata";
+        }
+
+        public string EliminaLezione(int idLezioneDaEliminare)
+        {
+            //controllo su input
+            var lezioneEsistente = lezioniRepo.GetById(idLezioneDaEliminare);
+            if (lezioneEsistente == null)
+            {
+                return "Id non valido.Impossibile eliminare.";
+            }
+            lezioniRepo.Delete(lezioneEsistente);
+            return "Lezione eliminata con successo";
+        }
+
+        public IList<Lezione> GetLezioniByCodiceCorso(string codiceCorso)
+        {
+            //TODO: controllare input
+            if (string.IsNullOrEmpty(codiceCorso) == true)
+            {
+                return null;
+            }
+            var lezioni = new List<Lezione>();
+            lezioni = lezioniRepo.GetAll().Where(l => l.CorsoCodice == codiceCorso).ToList();
+            return lezioni;
+        }
+
+        public IList<Lezione> GetLezioniByNomeCorso(string nomeCorso)
+        {
+            var corso = corsiRepo.GetAll().FirstOrDefault(c => c.Nome == nomeCorso);
+
+            if (corso == null)
+            {
+                return null;
+            }
+
+            var lezioni = new List<Lezione>();
+            lezioni = lezioniRepo.GetAll().Where(l => l.CorsoCodice == corso.CorsoCodice).ToList();
+            return lezioni;
         }
         #endregion
     }
